Extracted layout views into their own and some generic styles, to avoid repeating code. Also renamed a fiew drawable and layout files to fit the naming convention

+++ app/src/test/java/movienight/javi/com/movienight/MovieRequestTest.java
@@ -39,7 +39,7 @@ public class MovieRequestTest {
     public void setMovieRequestRating() throws Exception {
 
         // Arrange
-        double expectedRating = 6.5;
+        Float expectedRating = 6.5f;
 
         // Act
         mMovieRequest.setRating(expectedRating);
